Refactor JS code and change loading order of JS
JQuery and momentjs are now loaded before the page code because they
are dependencies.

The data is loaded as soon as possible to help parallelise server
requests.

A simple templating engine has been added so that we can reason about
the switch statement that builds the card HTML a little more easily.